should compile now.
Changed some of the variables so that the code should now compile in quartus, in the state table the next state wasn't a valid var, so added a temp state that will be fixed later, in the initializing of Datapath has extra variables that are being read so added temp input values that will be changed later. changed the rest of x_3,y_3 etc to current variables that we have and other mismatched variables.